backend/requirements.txt
django==5.2.13
djangorestframework==3.17.1
django-filter==25.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
pillow==10.4
whitenoise==6.7
gunicorn==22.0
psycopg2-binary==2.9.10
dj-database-url==2.2
cloudinary==1.40
django-cloudinary-storage==0.3
python-dotenv==1.0
django-extensions==3.2
setuptools==80.9.0
celery==5.4.0
redis==5.0.8
django-redis==5.4.0
